+++ app/Item.php
@@ -45,32 +45,6 @@ class Item extends Model implements IndexableModel, TranslatableContract
 
     protected $keyType = 'string';
 
-    public static $filterables = [
-        'author',
-        'topic',
-        'work_type',
-        'medium',
-        'technique',
-        'gallery',
-        'additionals.category.keyword',
-        'additionals.frontend.keyword',
-        'additionals.set.keyword',
-        'additionals.location.keyword',
-    ];
-
-    public static $rangeables = [
-        'date_earliest',
-        'date_latest',
-        'additionals.order',
-    ];
-
-    public static $sortables = [
-        'date_earliest',
-        'date_latest',
-        'view_count',
-        'additionals.order',
-    ];
-
     public $translatedAttributes = [
         'title',
         'description',
@@ -803,24 +777,5 @@ class Item extends Model implements IndexableModel, TranslatableContract
     public function getImageUrlAttribute()
     {
         return sprintf('%s%s', config('app.url'), $this->getImagePath());
-    }
-
-    public static function filterQuery(array $filter, BoolQueryBuilder $builder = null)
-    {
-        $builder = $builder ?: new BoolQueryBuilder();
-        foreach ($filter as $field => $value) {
-            if (is_string($value) && in_array($field, self::$filterables, true)) {
-                $builder->filter('term', [$field => $value]);
-            } else if (is_array($value) && in_array($field, self::$rangeables, true)) {
-                $range = collect($value)
-                    ->only(['lt', 'lte', 'gt', 'gte'])
-                    ->transform(function ($value) {
-                        return (string)$value;
-                    })
-                    ->all();
-                $builder->filter('range', [$field => $range]);
-            }
-        }
-        return $builder;
     }
 }